Moved cache to own file
